+++ ChangeLog.md
@@ -19,10 +19,23 @@ All notable changes to this project will be documented in this file.
 - SCSS structure
 - main SCSS
 - fonts
-- JSON server
 - parcel package for static files
 
 ### FIXED:
 
 - package.json
 - faq section
+
+## [0.2.0] - 23.01.2024
+
+### ADDED:
+
+- navigation
+- animations
+- create HamburgerMenu class
+- create Carousel class
+- create Questions class
+
+### FIXED:
+
+- styles